+++ src/databricks/labs/remorph/reconcile/runner.py
@@ -10,7 +10,7 @@ from databricks.sdk.errors import NotFound, PermissionDenied
 
 from databricks.labs.remorph.config import ReconcileConfig, TableRecon
 from databricks.labs.remorph.deployment.recon import RECON_JOB_NAME
-from databricks.labs.remorph.reconcile.execute import RECONCILE_OPERATION_NAME
+from databricks.labs.remorph.reconcile.recon_config import RECONCILE_OPERATION_NAME
 
 logger = logging.getLogger(__name__)
