+++ tests/integration/reconcile/query_builder/test_execute.py
@@ -29,7 +29,7 @@ from databricks.labs.remorph.reconcile.execute import (
     recon,
     generate_volume_path,
 )
-from databricks.labs.remorph.reconcile.recon_config import (
+from databricks.labs.remorph.reconcile.recon_output_config import (
     DataReconcileOutput,
     MismatchOutput,
     ThresholdOutput,
