+++ tests/unit/test_install.py
@@ -446,147 +446,6 @@ def test_configure_transpile_installation_with_no_validation(ws, ws_installer):
     )
 
 
-def test_configure_transpile_installation_with_validation_and_cluster_id_in_config(ws_installer, ws):
-    prompts = MockPrompts(
-        {
-            r"Select the source dialect": ALL_INSTALLED_DIALECTS.index("snowflake"),
-            r"Select the transpiler": TRANSPILERS_FOR_SNOWFLAKE.index("Morpheus"),
-            r"Enter input SQL path.*": "/tmp/queries/snow",
-            r"Enter output directory.*": "/tmp/queries/databricks",
-            r"Enter error file path.*": "/tmp/queries/errors.log",
-            r"Would you like to validate.*": "yes",
-            r"Do you want to use SQL Warehouse for validation?": "no",
-            r"Open .* in the browser?": "no",
-        }
-    )
-    installation = MockInstallation()
-    ws.config.cluster_id = "1234"
-
-    resource_configurator = create_autospec(ResourceConfigurator)
-    resource_configurator.prompt_for_catalog_setup.return_value = "remorph_test"
-    resource_configurator.prompt_for_schema_setup.return_value = "transpiler_test"
-
-    ctx = ApplicationContext(ws)
-    ctx.replace(
-        prompts=prompts,
-        installation=installation,
-        resource_configurator=resource_configurator,
-        workspace_installation=create_autospec(WorkspaceInstallation),
-    )
-
-    workspace_installer = ws_installer(
-        ctx.workspace_client,
-        ctx.prompts,
-        ctx.installation,
-        ctx.install_state,
-        ctx.product_info,
-        ctx.resource_configurator,
-        ctx.workspace_installation,
-    )
-
-    config = workspace_installer.configure(module="transpile")
-
-    expected_config = RemorphConfigs(
-        transpile=TranspileConfig(
-            transpiler_config_path=PATH_TO_TRANSPILER_CONFIG,
-            transpiler_options={},
-            source_dialect="snowflake",
-            input_source="/tmp/queries/snow",
-            output_folder="/tmp/queries/databricks",
-            error_file_path="/tmp/queries/errors.log",
-            catalog_name="remorph_test",
-            schema_name="transpiler_test",
-            sdk_config={"cluster_id": "1234"},
-        )
-    )
-    assert config == expected_config
-    installation.assert_file_written(
-        "config.yml",
-        {
-            "transpiler_config_path": PATH_TO_TRANSPILER_CONFIG,
-            "catalog_name": "remorph_test",
-            "input_source": "/tmp/queries/snow",
-            "output_folder": "/tmp/queries/databricks",
-            "error_file_path": "/tmp/queries/errors.log",
-            "schema_name": "transpiler_test",
-            "sdk_config": {"cluster_id": "1234"},
-            "source_dialect": "snowflake",
-            "version": 3,
-        },
-    )
-
-
-def test_configure_transpile_installation_with_validation_and_cluster_id_from_prompt(ws_installer, ws):
-    prompts = MockPrompts(
-        {
-            r"Select the source dialect": ALL_INSTALLED_DIALECTS.index("snowflake"),
-            r"Select the transpiler": TRANSPILERS_FOR_SNOWFLAKE.index("Morpheus"),
-            r"Enter input SQL path.*": "/tmp/queries/snow",
-            r"Enter output directory.*": "/tmp/queries/databricks",
-            r"Enter error file path.*": "/tmp/queries/errors.log",
-            r"Would you like to validate.*": "yes",
-            r"Do you want to use SQL Warehouse for validation?": "no",
-            r"Enter a valid cluster_id to proceed": "1234",
-            r"Open .* in the browser?": "no",
-        }
-    )
-    installation = MockInstallation()
-    ws.config.cluster_id = None
-
-    resource_configurator = create_autospec(ResourceConfigurator)
-    resource_configurator.prompt_for_catalog_setup.return_value = "remorph_test"
-    resource_configurator.prompt_for_schema_setup.return_value = "transpiler_test"
-
-    ctx = ApplicationContext(ws)
-    ctx.replace(
-        prompts=prompts,
-        installation=installation,
-        resource_configurator=resource_configurator,
-        workspace_installation=create_autospec(WorkspaceInstallation),
-    )
-
-    workspace_installer = ws_installer(
-        ctx.workspace_client,
-        ctx.prompts,
-        ctx.installation,
-        ctx.install_state,
-        ctx.product_info,
-        ctx.resource_configurator,
-        ctx.workspace_installation,
-    )
-
-    config = workspace_installer.configure(module="transpile")
-
-    expected_config = RemorphConfigs(
-        transpile=TranspileConfig(
-            transpiler_config_path=PATH_TO_TRANSPILER_CONFIG,
-            transpiler_options={},
-            source_dialect="snowflake",
-            input_source="/tmp/queries/snow",
-            output_folder="/tmp/queries/databricks",
-            error_file_path="/tmp/queries/errors.log",
-            catalog_name="remorph_test",
-            schema_name="transpiler_test",
-            sdk_config={"cluster_id": "1234"},
-        )
-    )
-    assert config == expected_config
-    installation.assert_file_written(
-        "config.yml",
-        {
-            "transpiler_config_path": PATH_TO_TRANSPILER_CONFIG,
-            "catalog_name": "remorph_test",
-            "input_source": "/tmp/queries/snow",
-            "output_folder": "/tmp/queries/databricks",
-            "error_file_path": "/tmp/queries/errors.log",
-            "schema_name": "transpiler_test",
-            "sdk_config": {"cluster_id": "1234"},
-            "source_dialect": "snowflake",
-            "version": 3,
-        },
-    )
-
-
 def test_configure_transpile_installation_with_validation_and_warehouse_id_from_prompt(ws_installer, ws):
     prompts = MockPrompts(
         {
